// src/types/getAddressTypes.ts
export type IAddress = {
    apartment: string;
    client_id: number;
    door: string;
    door_image: string | null;
    entrance: string;
    floor: string;
    has_elevator: number;
    home: string;
    id: number;
    is_apartment: number;
    latitude: string;
    longitude: string;
    name: string;
    point: string;
    region_id: number;
    street: string;
  };
  